+++ src/main/java/com/mclegoman/perspective/client/screen/config/shaders/ShadersConfigScreen.java
@@ -79,7 +79,7 @@ public class ShadersConfigScreen extends Screen {
         GRID.getMainPositioner().alignHorizontalCenter().margin(2);
         GridWidget.Adder GRID_ADDER = GRID.createAdder(2);
         GRID_ADDER.add(ButtonWidget.builder(Translation.getConfigTranslation("shaders.shader", new Object[]{ShaderDataLoader.getShaderName((int) ConfigHelper.getConfig("super_secret_settings"))}, new Formatting[]{Shader.getRandomColor()}), (button) -> {
-            if ((boolean) ConfigHelper.getExperimentalConfig("super_secret_settings_list")) ClientData.CLIENT.setScreen(new ShaderSelectionConfigScreen(new ShadersConfigScreen(PARENT_SCREEN, SAVE_ON_CLOSE, false)));
+            if ((boolean) ConfigHelper.getExperimentalConfig("super_secret_settings_list")) ClientData.CLIENT.setScreen(new ShaderSelectionConfigScreen(new ShadersConfigScreen(PARENT_SCREEN, SAVE_ON_CLOSE, false), -1));
             else {
                 Shader.cycle(ClientData.CLIENT, !REVERSE, true, false, false);
                 REFRESH = true;