+++ CHANGELOG.md
@@ -16,7 +16,8 @@ More information on this change will be posted as we get closer to 1.2.0-release
     - This option bypasses the panorama incompatibility list.  
   - The config version was updated to `12` in `1.2.0-beta.1`.  
 - **Depth Shaders will now render properly when an entity with glowing is rendered.**  
-  - If you experience any issues relating to this, try cycling your graphics options and report the issue.  
+  - If you experience any issues relating to this, try cycling your graphics options and report the issue.
+- The player's hand will now render over Depth Shaders.  
 - **Updated Take Panorama Screenshot.**  
   - The name format has been updated to match vanilla screenshots.  
   - Perspective Super Secret Settings will now be rendered on the panorama screenshots.